+++ Control/DeepSeq/BackDoor.hs
@@ -5,14 +5,6 @@
 -- This module only re-export reasonably safe entities from non-safe
 -- modules when there is no safe alternative
 
-#if MIN_VERSION_base(4,9,0) || (MIN_VERSION_base(4,6,0) && !MIN_VERSION_base(4,7,0))
-{-# LANGUAGE Safe #-}
-
-module Control.DeepSeq.BackDoor
-       {-# WARNING "This module is empty! Do not import me!" #-}
-       () where
-
-#else
 {-# LANGUAGE Trustworthy #-}
 
 module Control.DeepSeq.BackDoor
@@ -20,14 +12,15 @@ module Control.DeepSeq.BackDoor
     ) where
 
 #if !(MIN_VERSION_base(4,6,0))
--- not SAFE
+-- not safe, but Down is in Data.Ord starting with base-4.6
 import GHC.Exts as X ( Down(Down) )
 #endif
 
-#if MIN_VERSION_base(4,10,0)
--- Data.Type.Equality SAFE starting with base-4.10
+#if MIN_VERSION_base(4,9,0)
+-- Data.Type.Equality present from base-4.7, trustworthy starting with base-4.9
 #elif MIN_VERSION_base(4,7,0)
 import Data.Type.Equality as X ( (:~:) )
 #endif
 
-#endif
+-- present since at least base-4.0, maybe sooner?
+import GHC.Conc as X ( TVar(TVar) )