+++ changelog.md
@@ -26,6 +26,15 @@
   * Add `GNFData` for URec
     This will enable deriving NFData instances for unboxed types
 
+  * Add `NFData` instance for `UArray`
+    ([#50](https://github.com/haskell/deepseq/pull/50))
+
+  * Add `NFData` instance for `ForeignPtr`
+    ([#50](https://github.com/haskell/deepseq/pull/50))
+
+  * Add `NFData` instance for `TVar`
+    ([#50](https://github.com/haskell/deepseq/pull/50))
+
 ## 1.4.4.0 *Sep 2018*
 
   * Bundled with GHC 8.6.1